Rename Spring common testing packages
