+++ instrumentation/spring/spring-cloud-gateway/spring-cloud-gateway-2.2/testing/src/test/java/io/opentelemetry/instrumentation/spring/gateway/v2_2/Gateway22RouteMappingTest.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.instrumentation.spring.gateway.v2_2;
 
-import io.opentelemetry.instrumentation.spring.gateway.common.AbstractRouteMappingTest;
+import io.opentelemetry.instrumentation.spring.cloud.gateway.common.AbstractRouteMappingTest;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest(
