+++ instrumentation/spring/spring-cloud-gateway/spring-cloud-gateway-2.2/testing/src/test/java/io/opentelemetry/instrumentation/spring/gateway/v2_2/Gateway22TestApplication.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.instrumentation.spring.gateway.v2_2;
 
-import io.opentelemetry.instrumentation.spring.gateway.common.GatewayTestApplication;
+import io.opentelemetry.instrumentation.spring.cloud.gateway.common.GatewayTestApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
