+++ instrumentation/spring/spring-cloud-gateway/spring-cloud-gateway-webflux-4.3/testing/src/test/java/io/opentelemetry/instrumentation/spring/gateway/v4_3/Gateway43RouteMappingTest.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.instrumentation.spring.gateway.v4_3;
 
-import io.opentelemetry.instrumentation.spring.gateway.common.AbstractRouteMappingTest;
+import io.opentelemetry.instrumentation.spring.cloud.gateway.common.AbstractRouteMappingTest;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest(
