+++ instrumentation/spring/spring-webmvc/spring-webmvc-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/webmvc/v3_1/boot/SecurityConfig.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.javaagent.instrumentation.spring.webmvc.v3_1.boot;
 
-import io.opentelemetry.instrumentation.spring.webmvc.boot.SavingAuthenticationProvider;
+import io.opentelemetry.instrumentation.spring.webmvc.v3_1.boot.SavingAuthenticationProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
