+++ instrumentation/spring/spring-webmvc/spring-webmvc-3.1/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/webmvc/v3_1/filter/ServletFilterTest.java
@@ -8,8 +8,8 @@ package io.opentelemetry.javaagent.instrumentation.spring.webmvc.v3_1.filter;
 import static io.opentelemetry.instrumentation.testing.util.TestLatestDeps.testLatestDeps;
 
 import com.google.common.collect.ImmutableMap;
-import io.opentelemetry.instrumentation.spring.webmvc.filter.AbstractServletFilterTest;
-import io.opentelemetry.instrumentation.spring.webmvc.filter.FilteredAppConfig;
+import io.opentelemetry.instrumentation.spring.webmvc.v3_1.filter.AbstractServletFilterTest;
+import io.opentelemetry.instrumentation.spring.webmvc.v3_1.filter.FilteredAppConfig;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerTestOptions;
