+++ instrumentation/spring/spring-webmvc/spring-webmvc-6.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/webmvc/v6_0/filter/ServletFilterTest.java
@@ -14,8 +14,8 @@ import static io.opentelemetry.semconv.ExceptionAttributes.EXCEPTION_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.instrumentation.spring.webmvc.filter.AbstractServletFilterTest;
-import io.opentelemetry.instrumentation.spring.webmvc.filter.FilteredAppConfig;
+import io.opentelemetry.instrumentation.spring.webmvc.v3_1.filter.AbstractServletFilterTest;
+import io.opentelemetry.instrumentation.spring.webmvc.v3_1.filter.FilteredAppConfig;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerTestOptions;
