+++ instrumentation/spring/spring-webmvc/spring-webmvc-common-3.1/testing/src/main/java/io/opentelemetry/instrumentation/spring/webmvc/v3_1/boot/SavingAuthenticationProvider.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.spring.webmvc.boot;
+package io.opentelemetry.instrumentation.spring.webmvc.v3_1.boot;
 
 import java.util.ArrayList;
 import java.util.List;
