+++ instrumentation/spring/spring-webmvc/spring-webmvc-common-3.1/testing/src/main/java/io/opentelemetry/instrumentation/spring/webmvc/v3_1/filter/AbstractServletFilterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.spring.webmvc.filter;
+package io.opentelemetry.instrumentation.spring.webmvc.v3_1.filter;
 
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.CAPTURE_HEADERS;
 import static io.opentelemetry.instrumentation.testing.junit.http.ServerEndpoint.ERROR;
